resolve feedback for support function PR #860
